requirements.txt


# -------------------------
# Deep Learning
# -------------------------
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# -------------------------
# Data & Scientific Stack
# -------------------------
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
scipy>=1.9.0

# -------------------------
# Image Processing
# -------------------------
Pillow>=9.5.0
opencv-python>=4.8.0

# -------------------------
# Training Utilities
# -------------------------
tqdm>=4.65.0

# -------------------------
# Experiment Tracking
# -------------------------
mlflow>=2.9.0
wandb>=0.16.0

# -------------------------
# Explainability
# -------------------------
matplotlib>=3.7.0
seaborn>=0.12.0

# -------------------------
# RAG / LLM
# -------------------------
langchain>=0.1.0
faiss-cpu>=1.7.4
openai>=1.6.0
tiktoken>=0.5.0

# -------------------------
# API / Deployment
# -------------------------
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# -------------------------
# Utilities
# -------------------------
pyyaml>=6.0
requests>=2.31.0
